Move responsibility from benchmark to Timer
Timer now has the responsibility to keep record of all the laps.
It will also get a PerfVar, which is a performance counter, to count.
The performance counter will just be ignored by the non-Linux Timer,
but should be taken into account in the Linux version.

The reason for this is that apparently we can only count a single
value at a time, or the results will be weird. This extra argument
makes it easier to run auto-tests. We should be careful however to
run tests with the same random seed if we want to compare the graphs.

I'm tired...